move dual_stats_manager under airflow-core
