+++ providers/edge3/src/airflow/providers/edge3/worker_api/routes/worker.py
@@ -213,7 +213,7 @@ def set_state(
     worker.last_update = timezone.utcnow()
     session.commit()
     try:
-        from airflow._shared.observability.metrics.dual_stats_manager import DualStatsManager
+        from airflow.observability.metrics.dual_stats_manager import DualStatsManager
 
         DualStatsManager.incr(
             "edge_worker.heartbeat_count",
